requirements.txt
aiogram>=3.15.0
python-dotenv>=1.0.1
requests>=2.32.0
httpx>=0.27.0
PyYAML>=6.0.1
python-docx>=1.1.2
qrcode[pil]>=8.0
opencv-python-headless>=4.10.0.84
numpy>=1.26.0